+++ src/utils/helpers.ts
@@ -31,3 +31,8 @@ export const getCoinRecord = (
 
   return null;
 };
+
+export const formatWalletAddress = (address: string) => {
+  if (!address) return '';
+  return `${address.substring(0, 6)}...${address.substring(address.length - 4)}`;
+};
