+++ src/utils/wallet.ts
@@ -1,0 +1,16 @@
+import { PromotedWallet } from '../types';
+import metamaskLogo from '*.svg';
+import hotWalletLogo from '*.svg';
+
+export const PROMOTED_WALLETS: PromotedWallet[] = [
+  {
+    name: 'MetaMask',
+    logo: metamaskLogo,
+    downloadUrl: 'https://metamask.io/download/',
+  },
+  {
+    name: 'HOT Wallet',
+    logo: hotWalletLogo,
+    downloadUrl: 'https://hot-labs.org/extension/',
+  },
+];
